backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sentence-transformers
faiss-cpu
numpy
requests
python-dotenv
